requirements.txt
Django==1.9.6
django-extensions==1.6.7
django-filter==0.13.0
django-polymorphic==0.9.2
djangorestframework==3.3.3
factory-boy==2.7.0
ipython==4.2.0
pytest-django==2.9.1
pytest-sugar==0.7.1
pytest==2.9.2
python-dateutil==2.5.3
